data_scrapers/vistar/requirements.txt
requests >= 2.25.1
pytesseract >= 0.3.0
pillow >= 9.5.0
paho-mqtt >= 1.5.1